More cleanup and organization
